Set `minimum_deployment_os_version` for xcconfigs overrides
It set `minimum_deployment_os_version` given `SWIFT_DEPLOYMENT_TARGET`
or `LLVM_TARGET_TRIPLE_OS_VERSION` for `ios_application`.

Ideally, this may be handled by xcconfig, which would work for macOS as
well. We'd need a decent amount of changes to handle this in Bazel,
mainly incoporating targets and other aspects of xccconfig.

We'll need the `rules_ios` side of this as well:
https://github.com/bazel-ios/rules_ios/pull/267